requirements.txt
flask==2.3.3
flask-socketio==5.3.5
python-nmap==0.7.1
eventlet==0.33.3 